fix(app): keep first-run wizard across launch --project-dir
Snapshot the empty-store session count before applying --project-dir,
refuse symlink writes of orchestra-mode.json, and extract the wizard
DOM so main.ts stays within the desktop-split line budget.
