+++ .githooks/pre-commit-whitelist.ps1
@@ -189,6 +189,8 @@ $whitelistPatterns = @(
     'winsmux-app/src/workerPaneRouting.ts',
     'winsmux-app/src/workerPaneInput.ts',
     'winsmux-app/src/teamProfileSettings.ts',
+    'winsmux-app/src/firstRunOnboarding.ts',
+    'winsmux-app/src/firstRunWizard.ts',
     'winsmux-app/src/workflowGate.ts',
     'winsmux-app/src/styles.css',
     'winsmux-app/src-tauri/Cargo.toml',
